docs/architecture/interface-contracts.test.ts
/**
 * Test Specifications for Interface Contracts
 */

import {
  IGameEngine,
  INetworkManager,
  IStorageManager,
  IUIManager,
  IEventSystem
} from './interface-contracts';

describe('Interface Contract Tests', () => {
  // [Previous test cases remain the same...]

  describe('Performance Tests', () => {
    test('should handle high message volume', () => {
      // Test implementation
    });

    test('should manage memory efficiently', () => {
      // Test implementation
    });

    test('should scale with game size', () => {
      // Test implementation
    });
  });
});

// Test utilities
function createMockGameEngine(): IGameEngine {
  return {
    initializeGame: jest.fn(),
    processAction: jest.fn(),
    validateAction: jest.fn(),
    getGameState: jest.fn(),
    subscribeToEvents: jest.fn(),
    unsubscribeFromEvents: jest.fn(),
    createSnapshot: jest.fn(),
    restoreFromSnapshot: jest.fn()
  };
}

function createMockNetworkManager(): INetworkManager {
  return {
    connect: jest.fn(),
    disconnect: jest.fn(),
    getConnectionState: jest.fn(),
    sendMessage: jest.fn(),
    receiveMessage: jest.fn(),
    syncState: jest.fn(),
    requestSync: jest.fn()
  };
}

function createMockStorageManager(): IStorageManager {
  return {
    saveGameState: jest.fn(),
    loadGameState: jest.fn(),
    saveAction: jest.fn(),
    getActionHistory: jest.fn(),
    pruneHistory: jest.fn(),
    cleanup: jest.fn()
  };
}

function createMockUIManager(): IUIManager {
  return {
    renderGameState: jest.fn(),
    updateDisplay: jest.fn(),
    handleUserInput: jest.fn(),
    showFeedback: jest.fn(),
    queueAnimation: jest.fn(),
    cancelAnimation: jest.fn()
  };
}

function createMockEventSystem(): IEventSystem {
  return {
    emit: jest.fn(),
    on: jest.fn(),
    off: jest.fn(),
    processEvent: jest.fn(),
    filterEvents: jest.fn()
  };
}

// Test helpers
interface TestContext {
  gameEngine: IGameEngine;
  networkManager: INetworkManager;
  storageManager: IStorageManager;
  uiManager: IUIManager;
  eventSystem: IEventSystem;
}

function createTestContext(): TestContext {
  return {
    gameEngine: createMockGameEngine(),
    networkManager: createMockNetworkManager(),
    storageManager: createMockStorageManager(),
    uiManager: createMockUIManager(),
    eventSystem: createMockEventSystem()
  };
}

function createTestGameState() {
  return {
    id: 'test-game',
    players: new Map(),
    deck: [],
    pile: [],
    currentPlayer: '',
    status: 'INIT'
  };
}

function createTestAction() {
  return {
    type: 'PLAY_CARDS',
    playerId: 'test-player',
    cards: [],
    timestamp: Date.now()
  };
}

function createTestEvent() {
  return {
    type: 'CARDS_PLAYED',
    gameId: 'test-game',
    data: {},
    timestamp: Date.now()
  };
}

async function simulateNetworkDelay(ms: number = 100) {
  return new Promise(resolve => setTimeout(resolve, ms));
}

function assertStateConsistency(state: any) {
  expect(state).toBeDefined();
  expect(state.id).toBeDefined();
  expect(state.players).toBeDefined();
  expect(state.status).toBeDefined();
}

function assertEventOrder(events: any[]) {
  for (let i = 1; i < events.length; i++) {
    expect(events[i].timestamp).toBeGreaterThanOrEqual(events[i-1].timestamp);
  }
}

function assertValidAction(action: any) {
  expect(action).toBeDefined();
  expect(action.type).toBeDefined();
  expect(action.playerId).toBeDefined();
  expect(action.timestamp).toBeDefined();
}

export {
  createTestContext,
  createTestGameState,
  createTestAction,
  createTestEvent,
  simulateNetworkDelay,
  assertStateConsistency,
  assertEventOrder,
  assertValidAction
};